% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circulation}
\alias{circulation}
\title{circulation data set}
\format{
TBA
}
\source{
TBA
}
\description{
circulation data set
}
\examples{
#
}
\references{
See citation('earlywarnings')
}
\keyword{datasets}
