% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_ews.R
\name{find.optima}
\alias{find.optima}
\title{Description: find.optima}
\usage{
find.optima(f, detection.threshold = 0, bw, detection.limit = 1)
}
\arguments{
\item{f}{density}

\item{detection.threshold}{detection threshold for peaks}

\item{bw}{bandwidth}

\item{detection.limit}{Minimun accepted density for a maximum;
as a multiple of kernel height

Returns:
@return A list with the following elements:
min minima
max maxima
detection.density Minimum detection density}
}
\description{
Detect optima, excluding very local optima below detection.threshold
}
\details{
Arguments:
}
\examples{
#

}
\references{
See citation('TBA')
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
